calculator worked and printed value

+++ app/Models/ValuesParameters.php
@@ -28,6 +28,6 @@ class ValuesParameters extends Model
     protected $casts = [
         'id' => 'string',
         'ParametersID'=>'string',
-        'Value' => 'integer',
+        'Value' => 'float',
     ];
 }
